requirements.txt
pypdf
python-dotenv~=1.0.1

langchain
langchain-openai
langchain_community
langchain-experimental
langchain-qdrant
qdrant_client

openai~=1.61.1
sentence-transformers

numpy~=1.26.4
streamlit~=1.42.2
dspy~=2.6.2

unstructured~=0.16.23
pdfminer.six
pi_heif
unstructured_inference
pdf2image

transformers~=4.49.0